pyproject.toml
[project]
name = "openai-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "afaq5587", email = "afaqlatif082@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "openai-agents[litellm]>=0.0.13",
]

[project.scripts]
openai-sdk = "openai_sdk:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/openai_sdk"]
